Explain the connection pool error message better
The previous message was misleading (especially for Ops guys) when
diagnosing problems related to the database connection.

The message was suggesting that the connection cannot be obtained which
normally assumes the need to look at the database.

But this isn't the case as the connection could not be retrieved from
the application's internal connection pool.

The new message should make it more explicit and remove the confusion.