Add method to print a duration in words
I think a method to print a duration in words is really useful.
Because if you calculate 2 Time objects with each other, a duration will come out, and at the moment there's no function to print a duration in words.

Example:

You have a start and an end date of a ticket.
To calculate how long the ticket was "open" you have to do this:
end time - start time

And to print that duration calulated above, you'd have to write your own helper or hack a little bit with `time_ago_in_words`.
So why not add a `duration_in_words` function?