ActiveRecord::Relation#destroy_all destroys records in batches
To destroy in batches, we had to change the return type of #destoy_all.
Since a lot of people may depend on the return type (and it is
documented), we left a little transition period quirk.

Apps upgraded to 5.0 will get a deprecation warning, on the 4.2
behaviour. To switch to the destroy in batches the following option
should be set:

   config.active_record.destroy_all_in_batches = true

Setting it in an initializer may be a good idea, so people don't end up
with a conditional behaviour between environments, which may be a tricky
problem to catch.