Remove misleading wording around session loading
It is not true that sessions will not be loaded if they are not used in a controller action. Rails will load the session cookie and write a new session cookie on every request regardless of wether the session cookie is being used as part of the request or not.