Add caveat to number_to_currency docs [ci skip]
I've worked on a few applications that have gone through the
internationalization process and had issues because they were using
`number_to_currency`. The minute a user is allowed to change their
locale, they can change the price displayed on a page from 10 US dollars
to 10 Mexican Pesos, which is far from the same amount of money.

Unlike other helpers that rely on i18n, `number_to_currency` does not
produce equivalent results when the locale is changed.

As I've explained this to a few groups of developers now, I thought it
might make for a good caveat in the docs.