Fix that `WhereClause#merge` was removing non-conflicting binds
`WhereClause#merge` previously removed conflicting predicates by column
name, then removed any binds that matched the column name of a
removed predicate. This can be a bit overzealous, as it can remove
binds which happen to coincide with the column name of the conflicting
predicate but which relate to a different table altogether.[1]

When such a thing occurs, a database-dependent error like so is presented:

  ActiveRecord::StatementInvalid: PG::ProtocolViolation: ERROR:  bind message supplies 4 parameters, but prepared statement "a17" requires 5
  : SELECT  "clubs".* FROM "clubs" INNER JOIN "memberships" ON "clubs"."id" = "memberships"."club_id" WHERE "clubs"."active" = $1 AND "memberships"."member_id" = $2 AND "memberships"."active" = $3 AND "memberships"."favourite" = $4 LIMIT $5

In this case, because either the non-conflicting bind for
`"clubs"."active"` or `"memberships"."active"` were removed from the
binds, so all the binds are shifted over by one, leaving `$5` without a
value.

The alternative here is to remove only binds which coincide with the
index of the removed bound predicate. As bound predicates reference
bind variables by index, e.g. `$1`, the correspondence here seems to be
correct.

Note that this does not cover all possible Arel predicates, but there
is now symmetry between what predicates and binds are excluded -
those for `equality_nodes?`, with LHS column match and RHS bind match.

[1] e.g. in the case where models are paranoid-deleted and several
default scopes reference the same deletion column name, which is how
we ran into it