Don't extend CollectionProxy with klass.current_scope
This introduces `proxy_scope` as an alternative to `scope` for
merging into CollectionProxy. The only difference is that it merges
in `default_scoped` rather than `all`, so that it never picks up
and caches values from `current_scope`.

`current_scope` continues to be picked up via existing calles to
`scope` and `target_scope`.

I also switched from `merge` to `merge!` in `target_scope` as that
saves a `spawn` call, which is unnecessary as the
`association_relation` is newly created for each call.