Make `postgresql_version` public
This is useful to libraries that want to feature gate based on the
version of PostgreSQL the user is connected to. For instance, I want to
know if the user is connected to a version of Postgres that supports
concurrent materialized view refreshes. I could add that as a method on
the adapter as a PR, but rails has no need for this itself.

Rails is already using the postgresql_version for its own feature gating
and this makes that possible for other libraries.