use parent RAILS_ENV for rake system call
load_schema_if_pending! runs a rake task to prepare the database if
necessary, but the child task does not necessarily run in the same Rails
environment as its parent. This could result in attempting to prepare
the wrong database depending on how the database is configured.

This change passes the environment to the task as an environment
variable which ensures the environment names match.