stop reversing forwarded ips
The first IP address in `X-Forwarded-For` is the original client. Currently, the order of the IP addresses is reversed, so if multiple proxies have forwarded the request, `calculate_ip` will return the final proxy, rather than the original client. This commit stops reversing the order, so that the original client is returned rather than the final proxy.