Allow filtering of non-Global ID ActiveJob args
**Why**:
In some cases, a job might include sensitive information as one of the arguments, and there might not be an easy way to prevent those arguments from being passed.

For example, Devise sends the raw `reset_password_token` as an argument to their mailer. This results in the token being exposed in Active Job logs. Modifying the Devise code to prevent the token from being passed is not trivial.

**How**:
Provide an option to remove all non-Global ID arguments from the logs, by setting `Rails.application.config.active_job.log_all_arguments = false`.

When the option is set to `true` (default setting), the Active Job logs are formatted as they were before, i.e. all arguments are exposed. When set to `false`, all arguments that don't respond to `to_global_id` are removed. If the job is an `ActionMailer::DeliveryJob`, then the arguments also include the mailer class name, method, and `deliver_now`.