monkey patch `recycle!` on to controllers once
Instead of checking whether the class has recycle! or not, we can just
always add the method to all controller classes when the test harness is
loaded.  Technically this means that the controller test harness will
not work with controllers that do not inherit from AC::Metal, but then,
I'm not sure that is supported anyway.

Mixing in the module one will ensure that we don't break method caches,
and eliminates a runtime check so it should speed up tests (slightly).