expires_now -> always_revalidate || http_cache_never
The original documentation on expires_now was slightly misleading, as in most
contexts "cache" is the same thing as "store". In HTTP caching, this is not
the case, as something which is "not cached" may still be stored and revalidated
on each request. Some users may have though the expires_now method prevented
browsers and intermediate caches from storing responses, which it would not.

In addition, expires_now was a bit of a confusing name for adding a cache-control
directive of "no-cache". By renaming "expires_now" to "always_revalidate", we are
revealing the intention behind the method.

Also, provided is an "http_cache_never" method intended to be the anti-counterpart
to "http_cache_forever". It adds a "no-store" directive, which will prevent the
response from being stored *or* cached by any caches.