[close #23009] Limit key length
Mysql has a weird bug where it cannot index a string column of utf8mb4 if it is over a certain character limit. To get compatibility with msql we can add a limit to the key column. 191 characters is a very long key, it seems reasonable to limit across all adapters since using a longer key wouldn't be supported in mysql.

Thanks to @kamipo for the original PR and the test refactoring.

Conversation: https://github.com/rails/rails/pull/23009#issuecomment-171416629