Deprecate the use of mutable query methods on `CollectionProxy`
Mutating the `CollectionProxy` is problematic because the underlying object
will accept the modifications, then throw them away soon after, at the next
`spawn`.

For example, at present, we get this unexpected behavior:

>> collection_proxy.order!('body ASC').order_values
=> ['body ASC']

>> collection_proxy.order!('body ASC').order('post_id ASC').order_values
=> ['post_id ASC']

Mutation is also problematic because `CollectionProxy` is cached in
`CollectionAssociation@proxy`. This warning is a step toward removing support for
`CollectionProxy` mutation.

[Ben Woosley & Victor Kmita]