Add initial support for allowing an error on order or limit of queries being ignored in batches
add some documentation and add 4 tests regarding error vs. warning behavior

fix a typo when referring to the message

go back to default in tests so that ordering is not important. use a constant instead of method. fix assert_nothing_raised call. use self.klass to allow per class configuration

remove logger warn assets as that is tested elsewhere. pass error_on_ignore through find_each and find_in_batches also.

add blocks to the finds so that the code is actually executed

put the setting back to default in an ensure

Add a changelog entry