disable controller / view thread spawning in tests
Tests can (and do) access the database from the main thread.  In this
case they were starting a transaction, then making a request.  The
request would create a new thread, which would allocate a new database
connection.  Since the main thread started a transaction that contains
data that the new thread wants to see, the new thread would not see it
due to data visibility from transactions.  Spawning the new thread in
production is fine because middleware should not be doing database
manipulation similar to the test harness.  Before 603fe20c it was
possible to set the database connection id based on a thread local, but
603fe20c changes the connection lookup code to never look at the
"connection id" but only at the thread object itself.  Without that
indirection, we can't force threads to use the same connection pool as
another thread.

Fixes #23483