check `supports_view_paths?` at registration time
If we check the conditional at registration time, then we can avoid the
conditional at runtime.  This commit checks for view path support when
the handler is registered so that runtime calls to `find_dependencies`
won't need to check the conditional.  The idea is that
`register_tracker` is called only once where `find_dependencies` may be
called many times.