Ensure prepared statement caching still occurs with Adequate Record
In Rails 5, we're much more restrictive about when we do or don't cache
a prepared statement. In particular, we never cache when we are sending
an IN statement or a SQL string literal

However, in the case of Adequate Record, we are *always* sending a raw
SQL string, and we *always* want to cache the result.

Fixes #23507

/cc @tgxworld