Prevent dirty checks from affecting original after #dup
When a record is cloned via #dup, the `@original_raw_attributes` cache
must also be reset, else the same hash is referenced on both the
original and newly cloned object.  Further changes to the clone cause
attributes to show as dirty on the original.

This seems to have been introduced with in-place dirty checking in 4.2,
and master is unaffected due to refactorings and better encapsulation.