Log to stdout by default
Originally this logic was added by my PR https://github.com/rails/rails/pull/11060/files to only log to stdout in development mode. This was to prevent double logging when STDOUT was already being set as a logger. Since the introduction of the ability to detect logging to STDOUT in https://github.com/rails/rails/pull/22933 we can inspect the destination of logs directly, this brings better dev/prod parity with logging behavior.