Refactor AV <=> AP interdependence
This commit helps to alleviate coupling between the `ActionController`,
`AbstractController`, and `ActionView` rendering stack. There were
multiple private methods that were only declared so that way the could
be overriden, there was one case in ActionView where `super` was called,
and it was assumed that Action View was operating within the
Action Pack rendering stack. This is despite the fact that Action Pack
is only a development dependency of Action View.

As well, a few methods that were previously `:nodoc:`'d were made
private methods, but this isn't a breaking change, because methods with
a `:nodoc:` comment means that they are not public API, and no end users
should be depending on them.