The JSON API media type should only work wih a JSON API handler
Since the media type 'application/vnd.api+json' is a spec,
it is inappropriate to handle it with the JSON renderer.

This PR removes support for a JSON API media type.

I would recommend the media type be registered on its own as `jsonapi`
when a jsonapi Renderer and deserializer (Http::Parameters::DEFAULT_PARSERS) are added.

Is related to work in https://github.com/rails/rails/pull/21496