Enable to set 'accept-charset' for a form
Before this commit developers can not set 'accept-charset' for
a form generated by `#form_for` or `#form_tag`
(always 'accept-charset' is set to 'UTF-8').
But sometimes developers want to create forms whose character encoding
is not 'UTF-8' (e.g. 'Windows-31J').
This commit enable us to set 'accept-charset' if `enforce_utf8` is false
and a `:charset` option is given.