cache all subnodes in the digest cache
this is for backwards compatibility.  We should remove this after
5-0-stable is created because digest calculation should be amortized.
Caching digests of subtrees will speed up digests for other trees that
share the same children.  However, this will also increase memory, and
after the app warms up, then those subtrees will never be reused which
means that memory is wasted.