Constrain the number of unused connections kept in the pool
Just as we don't fill the pool with connections at boot, we generally
needn't maintain a full pool just because we once needed that many.

For maximum compatibility, this behaviour is disabled by default, by
setting the minimum to match the configured maximum. But by choosing to
configure a lower minimum, and setting a suitable reaping_frequency
(e.g., 10 minutes), you can now get a dynamic pool that keeps just what
it is likely to need in the near future.