Update docs to point to new update task.
This guides were pointing to this command `rails app:update`, which I
tried to run, but it didnt worked. I think the right command is `rails
rails:update` instead. Also thats the name of the rake task.

Also I removed the `Rake` word from the title, as we run it using
`rails` bin now.

cc @kaspth
[skip ci]