Refinement of our "are you missing a template or did you omit it on purpose?" heuristics
Narrows the "are you in a browser, viewing the page?" check to exclude
non-GET requests. Allows content-less APIs to use implicit responses
without having to set a fake request format.

This will need further attention. If you forget to redirect from a POST
to a GET, you'll get a 204 No Content response that browsers will
typically treat as… do nothing. It'll seem like the form just didn't
work and knowing where to start debugging is non-obvious.

On the flip side, redirecting from POST and others is the default, done
everywhere, so it's less likely to be removed or otherwise missed.

Alternatives are to do more explicit browser sniffing.

Ref #23827.