Added some useful configuration options into configuring.md [ci skip]
Added information about inserting middlewares with indexes. I think this information useful. We can use indexes also for `insert_after` but the behavior is not same and this is not the right place to give full information about all configuration options.
The configuration for I18n fallbacks is important for production environment. Especially using different fallbacks for different locales.