Improve Action Cable message encoding + decoding
- This centralizes all parsing and processing methods in
a `ActionCable::Connection::Utils` concern, which is then `include`d for
use within the library.
- All payload parsing/processing is done immediately in
  `ActionCable::Connection::Base#receive`
- `ActionCable::Connection::Subscriptions` once again uses `id_key` as
  the identifier for subscriptions within the global hash. This was only
changed in the first place to accomodate payload parsing/processing.

See #23649 and #24040 for previous discussion regarding this.

Likely still more work to be done with this, but this is a start at
trying to organize this logic in a way that is easier to understand.