removes unreachable code
This array literal cannot be reached. The previous begin either
returns to the caller via the explicit return in the ensure
block if all goes well, or else propagates whatever make the
begin block abort execution.

I have investigated the origin of this a bit. In the past the
ensure block didn't have a return call, see for example c08547d.
Later on the return was added in 4da4506, but the trailing
literal was left there.