no need to clear an unusued collection
If the code reaches that line new_constants is no longer needed.
We only need here to iterate over it to discard stuff and done.

Note that constant_watch_stack.new_constants returns a new
reference each time it is invoked, so that #clear call was not
cleaning state in some internal structure (which would have been
a bit dirty as well at this level of coupling).