Add config/initializers/time_zone_aware_types.rb
This PR integrates #15726, under which `Time` columns should support time
zone aware attributes.

In order for this to be the **default behavior** on brand new Rails 5.0 apps,
an initializer is required. Without it, even brand new Rails 5.0 apps would be
showing the following deprecation warning, which should be reserved to apps
_ported_ to Rails 5.0 from an older version:

```
Time columns will become time zone aware in Rails 5.1. This
still causes `String`s to be parsed as if they were in `Time.zone`,
and `Time`s to be converted to `Time.zone`.
To keep the old behavior, you must add the following to your initializer:
    config.active_record.time_zone_aware_types = [:datetime]
To silence this deprecation warning, add the following:
    config.active_record.time_zone_aware_types << :time
```

@sgrif Is the above correct?