Fail db rake tasks when the configuration misses the database name
Before this change a rake db task would silently succeed without doing anything. This can easily happen if the database.yml defines the configuration from environment variables and those variables are not set. It's rather hard to debug and running the task with `--trace` is not helpful either. This change raises an exception pointing to the cause of the problem.