Dont clean connections on Integration tests
First, the action of `clear_active_connections!` should not be in the
QueryCache executor. They have different responsabilites.

This brings `ConnectionManagement` class back, now as an executor
instead of Middleware, however it is only resposible for clearing the
connections on complete.

Also, this adds a flag, so we can turn off the clear action on
Integration tests, otherwise we can return connections with open
transactions back to the pool. [fixes #24491]