edit pass over the RDoc of Rails::Railtie [ci skip]
The pass touches a few minor things.

One important aspect is that the wording treats "railtie" as a
regular noun. This is so because Railtie is a constant and
generally the text refers to "classes that act as railties",
which is different.

Railties are so fundamental in Rails that deserve a noun,
like engine, class, array, application, etc. They have a
meaning per se, beyond constants.

Also, make clear an extension's railtie does not need to
be called "Railtie" or have "Railtie" in their class name.