Run integration tests through the app's executor
This ensures that the "unit of computation" is an entire test, instead
of the request. This allows us to skip, for example, connection pooling.
This ensures that the same connection is used for the entire test.

We should arguably do this for all tests, but this level is the only
place that we have an obvious lifecycle to hook into.

I think there's still an underlying problem to address, which is that in
general we shouldn't release a connection with an open transaction that
hasn't been poisoned back into the pool. However, that's a slightly
larger change that can be addressed later.

Fixes #24004 (though this may require an upstream change for RSpec)
Fixes #23989
Close #24491
Close #24599