restores code comments in String#blank? [ci skip]
When you come here without context, it is important to hightlight that
checking the predicate is worthwhile due to the observation that blank
strings are often empty. So you complicate the code (which has a cost
in terms of readability and aesthetics), but statistically makes sense.

Then, you also need to explain why the second operand is so convoluted.

Otherwise, you wonder why this line is written precisely this way. That
is what code comments are for.