Rails 4.2 don't have any possibility to use several values for where.
Sometimes when class has many options in one enum we need to doing
where by several of them. In Rails 5.0 this functionality already
implemented but a lot users has project on Rails 4.2 and they can't
migrate them to Rails 5.0 right now but in any case they need to
search be several enum values. Queries with default enum abilities
became quite big and do the same. Also using int values is not
comfortable because if we have for example ten values, but for search
we need only values with indexes 1, 3, 4 we should always remember
what this values exactly means.

    class Book
      enum status: [:proposed, :written, :published, :reviewed]
    end

Previously where query for this code looks like:

    Book.where(status: [Book.statuses[:proposed], Book.statuses[:published]])

Now same query will be looks like:
    Book.where(status: Book.statuses_by_sym(:proposed, :published))

Much shorted and do the same!