Simplify AR::Migration and Migrator to use same connection logic
Previously, AR::Migration and Migrator were both dependent, and had
hardcoded, `ActiveRecord::Base.connection` as the connection object to
the database. Now, end users simplify have to override
`AR::Migration#connection` (not the instance method, the class method)
to be able to change the database connection that is used in this
section of the Active Record codebase.