Improve fixture load order
Currently, HABTM and has_many :through join table fixtures are loaded before
the fixtures for the parent table. This is because in the table_rows method,
the add_join_records method is called inside the map block that adds the
parent table fixtures to the rows hash. By initializing the parent table's key
to an empty array, we can ensure that the parent table will be loaded first.
The set of fixtures loaded will never be affected by this change, only the
ordering of the keys.