Fix Railties test
This test was failing due to Bundler v1.12.0 being released, which
exposed a flaw in how the following test from
`railties/test/application/initializers/frameworks_test.rb` was being
run:

```ruby
test "database middleware doesn't initialize when activerecord is not in frameworks" do
  use_frameworks []
  require "#{app_path}/config/environment"
  assert_nil defined?(ActiveRecord::Base)
end
```

Despite that the test assumed that Active Record was not loaded, it was.
I am honestly not sure how this test was passing before, but the release
of the previously mentioned version of Bundler broke this test
(rightfully so). Now, frameworks which should not be included (as this
test assumes) will not be loaded into the Rails application, and the
test will pass correctly.

This is a partial revert of #24839.