Try to coerce PostgreSQL array single elements types
Currently, the deserialization representation of a single element in an
array is always a string. I have hit this problem, while having a single
element deserialize method that expected an Integer, yet it got a
string.

Now, I'm not sure whether the Type#deserialize(value) do expects the
value to be coerced by the DB to a suitable type (or it should always
expect a string), but if the type is used standalone and not wrapped in
an array, I do get the coerced value.

This change is an effort in trying to hint the builtin
PG::Text{En,De}coders what the single elements type of the PostgreSQL
array is. If the type has no coder associated with it, we use the
stringified representation as before.