Provide a default CSP that is lenient yet secure
Building on the concept of default headers, it would be great if CSP was applied by default. This is mostly geared towards new applications. Unfortunately, making this "report only" will generate a warning in the console since the policy, understandably, lacks a report-uri value. Adding a default report-uri realistically would require adding a CSP reporting endpoint to rails by default. Because of these factors, applying the potentially breaking header probably better suited as part of a generator script for new applications but I don't know the cleanest way of accomplishing this.

In past lives (e.g. Twitter scala applications), applying CSP by default had amazing returns. The suggested policy should work for most applications and the browser console literally tells you what modifications need to be made. Adding code that violates this policy could be caught early on and people can choose to work within the suggested boundaries or they can decide to modify the policy. Naturally, there will be those who disable the policy entirely. That is entirely up to them but at least some thought will be put into it.

This is a bit contentious, of course.