Somewhat reduce AV <-- --> AP test module coupling
So, this is trying to make the best out of a terrible situation. Action
View (and its own tests (!!)), depend on including modules that provide
some testing API from all across Action Pack (AbstractController,
ActionDispatch, etc.) Without breaking a ton of Action View tests, and
creating an "enrichment API" similar to what we have going on with the
MIME code right now, I've just nested this in an if-constant-defined if
statement.

If there's a better way to do this, please let me know. :-(