Demonstration of logger missing Rack::Cache
One thing wrong with this. Rack::Cache is never actually
activated here so we're not really executing the full test.
But, I *think* this demonstrates that the logger does not respect
Rack::Cache. I know that this doesn't happen in our app.

Unfortunately, I'm not sure what the fix is here. If people have ideas,
I am happy to finish this pull request with the solution, but I'm
worried that there may be an architectural problem with how Rack::Cache
modifies the response too "deep" in the rack middleware stack for the
log subscriber to be aware of it.