Fix `rake db:migrate:status` with subdirectories
`db:migrate` supports subdirectories and have a test.

https://github.com/rails/rails/blob/v5.1.0.beta1/activerecord/test/cases/migrator_test.rb#L78-L85

But `db:migrate:status` doesn't work with subdirectories. It is due to
`Dir.foreach(path)` is not the same with `Dir["#{path}/**/[0-9]*_*.rb"]`.

I extracted `migration_files` and sharing it in the both to fix the
issue. And added tests for `db:migrate:status`.