bump mail version locking to ~> 2.6
mail < 2.6.0 includes a vulnerability. I'm not sure what the general policy is here or if there's a better locking strategy or what the potential breakage possibilities are.

The writeup for the issue is below:

The Mail library does not impose a length limit on email addresses, so an attacker can send a long spam message via a recipient address unless there is a limit on the application’s side. The attacker-injected message in the recipient address is processed by the server. This type of vulnerability can be real threats in inquiry forms, member signup forms, or any other application that delivers an email to a user-specified email address.

Affected versions: All versions prior to 2.6.0
Fixed versions: 2.6.0
Solution: Upgrade to latest version.
Credit: Takeshi Terada / Mitsui Bussan Secure Directions, Inc.
Sources: http://www.mbsd.jp/Whitepaper/smtpi.pdf
https://github.com/mikel/mail/commit/72befdc4dab3e6e288ce226a7da2aa474cf5be83