Improve error message when record touch fails.
The current error message only indicates that a touch can fail because the record is new. In practice, we saw cases where touches were failing because the record had been destroyed. `persisted?` checks `new_record?` *and* `destroyed?`. It was confusing to get a message about a new record when in reality we were destroying records.

I also included a helpful tip for users to consider using `persisted?`, `new_record?`, or `destroyed?` before touching.