Exclude 'tempfile' from UploadedFile's hash representation.
This patch adds a `#to_hash` to `ActionDispatch::Http:UploadedFile`
that excludes its `tempfile` instance variable from the hash
representation.

The latter is for example used while generating JSON from an
`UploadedFile` which leads to the *content* of the temporary file
being read and included in the JSON. This behaviour is most likely
not what a user would expect. Besides, it has been reported to result
in encoding conversion errors when the temporary file is a binary file
(`Tempfile#read` returns an ASCII-encoded String containing binary data
which `JSON` tries to `encode` that to UTF-8).

Fixes #25250.