Speed up RDoc generation when it's already been done once
Only include files that contain changes since the last generation
of the API documentation.

Thus, only modified files are re-generated instead of the whole API.

The `created.rid` file won't contain the whole list of files anymore
if we generate several times the API but we don't really care about it,
only the generation date and time is important.

It is possible to fall back to the previous behavior by defining the ALL
environment variable running the `rake rdoc` task.