Only use provided units for number_to_human
Fixes #25664

Previous implementation attempted to round numbers past the provided units. For
instance, if the number to be formatted was 1_000_000 and the highest custom
unit exponent was "thousand", the number returned would be reduced to 1, but
with a value for "million" not provided, no unit would be returned along with
the number. This change does not attempt to reduce the number farther than the
provided units.