reestablish current environment connection after creating all databases
Using `db:create:all` with `db:drop:all`, connection pool is last to have
information of the processing was carried out environment is held,
it may not match the environment for now processing.
In case of default `databases.yml`, it is probably left connection pool for
production environment.

Therefore, instead of reusing the connection pool, it fixes to perform reestablish
for current environment.

Fixes #24813